+++ src/client.rs
@@ -46,7 +46,7 @@ impl Client {
     /// (like unreachable host) so it's important that you handle those
     /// errors.
     pub fn get_connection(&self) -> RedisResult<Connection> {
-        Ok(connect(&self.connection_info, None)?)
+        connect(&self.connection_info, None)
     }
 
     /// Instructs the client to actually connect to redis with specified
@@ -55,7 +55,7 @@ impl Client {
     /// a variety of errors (like unreachable host) so it's important
     /// that you handle those errors.
     pub fn get_connection_with_timeout(&self, timeout: Duration) -> RedisResult<Connection> {
-        Ok(connect(&self.connection_info, Some(timeout))?)
+        connect(&self.connection_info, Some(timeout))
     }
 
     /// Returns a reference of client connection info object.
@@ -85,7 +85,7 @@ impl Client {
             }
         };
 
-        crate::aio::Connection::new(&self.connection_info, con).await
+        crate::aio::Connection::new(&self.connection_info.redis, con).await
     }
 
     /// Returns an async connection from the client.
@@ -235,7 +235,7 @@ impl Client {
         T: crate::aio::RedisRuntime,
     {
         let con = self.get_simple_async_connection::<T>().await?;
-        crate::aio::MultiplexedConnection::new(&self.connection_info, con).await
+        crate::aio::MultiplexedConnection::new(&self.connection_info.redis, con).await
     }
 
     async fn get_simple_async_connection<T>(
@@ -274,7 +274,7 @@ impl ConnectionLike for Client {
     }
 
     fn get_db(&self) -> i64 {
-        self.connection_info.db
+        self.connection_info.redis.db
     }
 
     fn check_connection(&mut self) -> bool {