+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+<a name="0.21.0"></a>
+### 0.20.0 (2021-07-16)
+
+
+#### Performance
+
+*   Don't enqueue multiplexed commands if the receiver is dropped ([ca5019db](https://github.com/mitsuhiko/redis-rs/commit/ca5019dbe76cc56c93eaecb5721de8fcf74d1641))
+
+#### Features
+
+*   Refactor ConnectionAddr to remove boxing and clarify fields
+
+<a name="0.20.2"></a>
+### 0.20.2 (2021-06-17)
+
+#### Features
+
+*   Provide a new_async_std function ([c3716d15](https://github.com/mitsuhiko/redis-rs/commit/c3716d154f067b71acdd5bd927e118305cd0830b))
+
+#### Bug Fixes
+
+*   Return Ready(Ok(())) when we have flushed all messages ([ca319c06](https://github.com/mitsuhiko/redis-rs/commit/ca319c06ad80fc37f1f701aecebbd5dabb0dceb0))
+*   Don't loop forever on shutdown of the multiplexed connection ([ddecce9e](https://github.com/mitsuhiko/redis-rs/commit/ddecce9e10b8ab626f41409aae289d62b4fb74be))
+
+
+
+<a name="0.20.1"></a>
+### 0.20.1 (2021-05-18)
+
+
+#### Bug Fixes
+
+*   Error properly if eof is reached in the decoder ([306797c3](https://github.com/mitsuhiko/redis-rs/commit/306797c3c55ab24e0a29b6517356af794731d326))
+
+
+
 <a name="0.20.0"></a>
 ## 0.20.0 (2021-02-17)